+++ admin/templates/info/setup.php
@@ -15,13 +15,13 @@ if ( ! defined( 'ABSPATH' ) ) {
 		<div class="awsm-job-setup-in">
 			<div class="awsm-job-setup-head">
 				<h1>
-				<a href="https://wpjobopenings.com" target="_blank">
+				<a href="<?php echo esc_url( AWSM_JOBS_SITE_URL ); ?>" target="_blank">
 						<img src="<?php echo esc_url( AWSM_JOBS_PLUGIN_URL . '/assets/img/hirezoot-logo-b.svg' ); ?>" align="<?php esc_html_e( 'HireZoot by AWSM INNOVATIONS', 'wp-job-openings' ); ?>">
 						<span><?php esc_html_e( 'HireZoot by AWSM INNOVATIONS', 'wp-job-openings' ); ?></span>
 					</a>
 				</h1>
 				<p>
-					<span><?php esc_html_e( 'Thanks for installing Hirezoot, you are awesome! ', 'wp-job-openings' ); ?></span>
+					<span><?php esc_html_e( 'Thanks for installing HireZoot, you are awesome! ', 'wp-job-openings' ); ?></span>
 					<span><?php esc_html_e( 'Your are a few minutes away from setting up your job page and start hiring.', 'wp-job-openings' ); ?></span>
 				</p>
 			</div><!-- .awsm-job-setup-head-->