+++ blocks/src/index.js
@@ -29,7 +29,7 @@ import { __ } from '@wordpress/i18n';
  * @see https://developer.wordpress.org/block-editor/reference-guides/block-api/block-registration/
  */
 registerBlockType( metadata.name, {
-	title: __( 'Job Listings (Beta)', 'wp-job-openings' ), // Block title.
+	title: __( 'Job Listings', 'wp-job-openings' ), // Block title.
 	description: __(
 		'Add and Customise your Job Listing Layout.',
 		'wp-job-openings'