+++ inc/templates/block-files/block-job-openings-view.php
@@ -83,6 +83,10 @@ $block_style_variables = "
 		--hz-pagination-border-radius-topright: {$styles['pagination_border_radius_topright']};
 		--hz-pagination-border-radius-bottomright: {$styles['pagination_border_radius_bottomright']};
 		--hz-pagination-border-radius-bottomleft: {$styles['pagination_border_radius_bottomleft']};
+		--hz-pagination-padding-top: {$styles['padding_top_pagination']};
+		--hz-pagination-padding-right: {$styles['padding_right_pagination']};
+		--hz-pagination-padding-bottom: {$styles['padding_bottom_pagination']};
+		--hz-pagination-padding-left: {$styles['padding_left_pagination']};
 
 		" . ( ! empty( $styles['sf_background_color'] ) ? "--hz-sf-bg-color: {$styles['sf_background_color']};" : '' ) . '
 		' . ( ! empty( $styles['sf_text_color'] ) ? "--hz-sf-tx-color: {$styles['sf_text_color']};" : '' ) . '
@@ -125,6 +129,21 @@ $has_alerts = (
 if ( $has_search || $has_filters || $has_alerts ) {
 	$show_filter             = true;
 	$placement_sidebar_class = 'awsm-job-2-col';
+}
+
+$button_style        = ! empty( $styles['button_style'] ) ? $styles['button_style'] : 'none';
+$button_style_filter = function ( $class ) use ( $button_style ) {
+	return $class . ' is-button-' . $button_style;
+};
+add_filter( 'awsm_b_job_more_button_class', $button_style_filter );
+
+$button_text        = ! empty( $styles['button_text'] ) ? $styles['button_text'] : '';
+$button_text_filter = null;
+if ( $button_text ) {
+	$button_text_filter = function () use ( $button_text ) {
+		return esc_html( $button_text );
+	};
+	add_filter( 'awsm_b_job_more_button_text', $button_text_filter );
 }
 
 $wrapper_class = 'awsm-b-job-wrap' . awsm_jobs_wrapper_class( false );
@@ -242,3 +261,8 @@ if ( $placement === 'top' ) {
 				<?php awsm_block_jobs_load_more( $query, $attributes ); ?>
 			</div>
 		</div>
+<?php
+remove_filter( 'awsm_b_job_more_button_class', $button_style_filter );
+if ( $button_text_filter ) {
+	remove_filter( 'awsm_b_job_more_button_text', $button_text_filter );
+}