+++ inc/widgets/class-awsm-job-openings-dashboard-widget.php
@@ -20,7 +20,7 @@ class AWSM_Job_Openings_Dashboard_Widget {
 
 	public function dashboard_setup() {
 		if ( current_user_can( 'edit_jobs' ) ) {
-			wp_add_dashboard_widget( 'awsm-jobs-overview-dashboard', esc_html__( 'Hirezoot - Overview', 'wp-job-openings' ), array( $this, 'display_widget' ) );
+			wp_add_dashboard_widget( 'awsm-jobs-overview-dashboard', esc_html__( 'HireZoot - Overview', 'wp-job-openings' ), array( $this, 'display_widget' ) );
 		}
 	}